Removed numpy dep

+++ robotiq_modbus_controller/request.py
@@ -1,3 +1,17 @@
+# Copyright 2021 Norwegian University of Science and Technology.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from dataclasses import dataclass
 import numpy as np
 
@@ -24,7 +38,7 @@ class GripperOptions:
 
 @dataclass
 class PositionRequest:
-    pr: np.uint8 = 0
+    pr: int = 0
 
     def to_byte(self):
         return self.pr
@@ -32,7 +46,7 @@ class PositionRequest:
 
 @dataclass
 class Speed:
-    sp: np.uint8 = 0
+    sp: int = 0
 
     def to_byte(self):
         return self.sp
@@ -40,7 +54,7 @@ class Speed:
 
 @dataclass
 class Force:
-    fr: np.uint8 = 0
+    fr: int = 0
 
     def to_byte(self):
         return self.fr
