+++ robotiq_modbus_controller/status.py
@@ -1,3 +1,17 @@
+# Copyright 2021 Norwegian University of Science and Technology.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from dataclasses import dataclass
 from typing import List
 import struct
@@ -8,29 +22,29 @@ import numpy as np
 class GripperStatus:
     act: bool
     gto: bool
-    sta: np.uint8
-    obj: np.uint8
+    sta: int
+    obj: int
 
 
 @dataclass
 class FaultStatus:
-    flt: np.uint8
-    kflt: np.uint8
+    flt: int
+    kflt: int
 
 
 @dataclass
 class PositionRequestEcho:
-    pr: np.uint8
+    pr: int
 
 
 @dataclass
 class Position:
-    po: np.uint8
+    po: int
 
 
 @dataclass
 class Current:
-    cu: np.uint8
+    cu: int
 
 
 @dataclass
